+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [Feature] Results parse HTTP URLs to make them clickable (#230)
 * [Feature] Adds `ggplot` plotting API (histogram and boxplot)
 * [Fix] Clearer error when using `--with` with snippets that do not exist (#257)
+* [Fix] Pytds now automatically compatible
 ## 0.6.6 (2023-03-16)
 
 * [Fix] Pinning SQLAlchemy 1.x